Fixed issue with Pandas rounding float16

+++ data.py
@@ -145,9 +145,9 @@ def generate_and_export_content_features():
     test_df = combine_features(rating_test_df, titles_df, mean_rating_df, genres_df, user_mean_rating_df)
 
     input_data_cols = MOVIE_FEATURES + USER_FEATURES + RATING_FEATURE
-    train_df[input_data_cols] = train_df[input_data_cols].round(2).astype(np.float16)
-    dev_df[input_data_cols] = dev_df[input_data_cols].round(2).astype(np.float16)
-    test_df[input_data_cols] = test_df[input_data_cols].round(2).astype(np.float16)
+    train_df[input_data_cols] = train_df[input_data_cols].astype(np.float32).round(2)
+    dev_df[input_data_cols] = dev_df[input_data_cols].astype(np.float32).round(2)
+    test_df[input_data_cols] = test_df[input_data_cols].astype(np.float32).round(2)
 
     export_to_parquet(train_df[MOVIE_FEATURES], train_movie_features_path)
     export_to_parquet(train_df[USER_FEATURES], train_user_features_path)
